Move ebulletin logic into About controller
