+++ controllers/toplevel/index.js
@@ -1,10 +1,8 @@
 'use strict';
 
-const Raven = require('raven');
 const express = require('express');
 const config = require('config');
 const { sortBy } = require('lodash');
-const rp = require('request-promise-native');
 const { body, validationResult } = require('express-validator/check');
 const xss = require('xss');
 const moment = require('moment');
@@ -15,8 +13,7 @@ const appData = require('../../modules/appData');
 const contentApi = require('../../services/content-api');
 const surveyService = require('../../services/surveys');
 const routeStatic = require('../utils/routeStatic');
-const getSecret = require('../../modules/get-secret');
-const analytics = require('../../modules/analytics');
+
 const { heroImages } = require('../../modules/images');
 const regions = require('../../config/content/regions.json');
 const addSection = require('../../middleware/addSection');
@@ -74,126 +71,6 @@ module.exports = (pages, sectionPath, sectionId) => {
 
     // Handle all the proxied legacy pages
     legacyPages.init(router);
-
-    // send form data to the (third party) email newsletter provider
-    // @TODO translate these error messages
-    router.post(
-        '/ebulletin',
-        [
-            body('firstName', 'Please provide your first name')
-                .exists()
-                .not()
-                .isEmpty(),
-            body('lastName', 'Please provide your last name')
-                .exists()
-                .not()
-                .isEmpty(),
-            body('email')
-                .exists()
-                .not()
-                .isEmpty()
-                .withMessage('Please provide your email address')
-                .isEmail()
-                .withMessage('Please provide a valid email address'),
-            body('location', 'Please choose a country newsletter')
-                .exists()
-                .not()
-                .isEmpty()
-        ],
-        (req, res) => {
-            const errors = validationResult(req);
-            // sanitise input
-            for (let key in req.body) {
-                req.body[key] = xss(req.body[key]);
-            }
-
-            if (!errors.isEmpty()) {
-                req.flash('formErrors', errors.array());
-                req.flash('formValues', req.body);
-                req.session.save(() => {
-                    res.redirect('/#' + config.get('anchors.ebulletin'));
-                });
-            } else {
-                let newsletterLocation = req.body.location;
-                let locale = req.body.locale;
-                let localePrefix = locale === 'cy' ? config.get('i18n.urlPrefix.cy') : '';
-
-                // redirect errors back to the homepage
-                let handleSignupError = errMsg => {
-                    Raven.captureMessage(errMsg || 'Error with ebulletin');
-                    req.flash('ebulletinStatus', 'error');
-                    req.session.save(() => {
-                        // @TODO build this URL more intelligently
-                        return res.redirect(localePrefix + '/#' + config.get('anchors.ebulletin'));
-                    });
-                };
-
-                let handleSignupSuccess = () => {
-                    analytics.track('emailNewsletter', 'signup', newsletterLocation);
-                    req.flash('ebulletinStatus', 'success');
-                    req.session.save(() => {
-                        // @TODO build this URL more intelligently
-                        return res.redirect(localePrefix + '/#' + config.get('anchors.ebulletin'));
-                    });
-                };
-
-                let dataToSend = {
-                    email: req.body.email,
-                    emailType: 'Html',
-                    dataFields: [
-                        {
-                            key: 'FIRSTNAME',
-                            value: req.body.firstName
-                        },
-                        {
-                            key: 'LASTNAME',
-                            value: req.body.lastName
-                        },
-                        {
-                            key: newsletterLocation,
-                            value: 'yes'
-                        }
-                    ]
-                };
-
-                // optional fields
-                if (req.body['organisation']) {
-                    dataToSend.dataFields.push({
-                        key: 'ORGANISATION',
-                        value: req.body.organisation
-                    });
-                }
-
-                let addressBookId = 589755;
-                let apiAddContactPath = `/address-books/${addressBookId}/contacts`;
-
-                // send the valid form to the signup endpoint (external)
-                rp({
-                    uri: config.get('ebulletinApiEndpoint') + apiAddContactPath,
-                    method: 'POST',
-                    auth: {
-                        user: getSecret('dotmailer.api.user'),
-                        pass: getSecret('dotmailer.api.password'),
-                        sendImmediately: true
-                    },
-                    json: true,
-                    body: dataToSend,
-                    resolveWithFullResponse: true
-                })
-                    .then(response => {
-                        // signup succeeded
-                        if (response.statusCode === 200) {
-                            return handleSignupSuccess();
-                        } else {
-                            return handleSignupError(response.message);
-                        }
-                    })
-                    .catch(error => {
-                        return handleSignupError(error.message || error);
-                    });
-            }
-        }
-    );
 
     // data page
     router.get(pages.data.path, (req, res) => {
